+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [BREAKING] Renamed public accessors of the `Block` struct to match the updated fields (#791).
 - [BREAKING] Changed the `TransactionArgs` to use `AdviceInputs` (#793).
 - Setters in `memory` module don't drop the setting `Word` anymore (#795).
+- [BREAKING] Increase of nonce does not require changes in account state any more (#789).
 
 ## 0.4.0 (2024-07-03)